+++ rules/plugins/chai-expect.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-	rules: {
-		'chai-expect/no-inner-compare': 'error',
-		'chai-expect/missing-assertion': 'error',
-		'chai-expect/terminating-properties': 'error',
-	},
-};