+++ src/config/mocha.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const includeRules = require('../lib/include-rules');
+
+module.exports = (configs, options) => ({
+	plugins: [
+		'mocha',
+	],
+	env: {
+		mocha: true,
+	},
+	rules: {
+		...includeRules('plugins/mocha', configs, options),
+	},
+});