+++ src/lib/include-rules.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const path = require('path');
+
+module.exports = (category, configs, options) => {
+	// eslint-disable-next-line security/detect-non-literal-require
+	const rules = require(path.resolve(__dirname, '..', 'rules', `${category}.js`));
+	return rules(configs, options);
+};