+++ src/rules/ecmascript-8.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = () => ({
+	'promise/prefer-await-to-then': 'error',
+	'promise/prefer-await-to-callbacks': 'error',
+});