+++ src/rules/plugins/chai-expect.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = () => ({
+	'chai-expect/no-inner-compare': 'error',
+	'chai-expect/missing-assertion': 'error',
+	'chai-expect/terminating-properties': 'error',
+});