+++ src/rules/plugins/node.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const bestPractices = require('./node/best-practices');
+const possibleErrors = require('./node/possible-errors');
+const stylisticIssues = require('./node/stylistic-issues');
+
+module.exports = (configs, options) => ({
+	...bestPractices(configs, options),
+	...possibleErrors(configs, options),
+	...stylisticIssues(configs, options),
+});