+++ src/rules/plugins/node/best-practices.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = () => ({
+	'node/no-deprecated-api': 'error',
+});